Fixes MySQL Database connection issue
An issue exisst with the last release of Barnyard that the app will close
if the default connection timeout limit is hit when trying to talk to the
database. This fixes a variable type that allows a larger value and adds
additional logging. This is useful when you add the following text to the
end of the 'output database' line in barnyard2.conf:
connection_limit=1000000

Please enter the commit message for your changes. Lines starting